backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow==12.0.0
djangorestframework-simplejwt==5.5.1

